requirements.txt
# Core API dependencies
openai>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
aiofiles>=23.2.0

# Vector database
pinecone-client>=3.0.0

# Database drivers (optional - install based on your DB type)
# For SQLite: built-in, no package needed
# For SQL Server: uncomment the appropriate line for your platform
# pyodbc>=4.0.39; sys_platform == 'win32'
# pyodbc>=4.0.39; sys_platform == 'linux'
# pyodbc>=4.0.39; sys_platform == 'darwin'

# UI dependencies (optional - only if using Streamlit UI)
streamlit>=1.28.0
requests>=2.31.0

